+++ src/components/scenes/TransactionDetailsScene.tsx
@@ -1,3 +1,4 @@
+import { useQuery } from '@tanstack/react-query'
 import { abs } from 'biggystring'
 import {
   EdgeAccount,
@@ -9,6 +10,7 @@ import {
   EdgeTxSwap
 } from 'edge-core-js'
 import * as React from 'react'
+import { useMemo } from 'react'
 import { View } from 'react-native'
 import FastImage from 'react-native-fast-image'
 import IonIcon from 'react-native-vector-icons/Ionicons'
@@ -34,8 +36,10 @@ import { getExchangeDenom } from '../../selectors/DenominationSelectors'
 import { useSelector } from '../../types/reactRedux'
 import { EdgeAppSceneProps } from '../../types/routerTypes'
 import { getCurrencyCodeWithAccount } from '../../util/CurrencyInfoHelpers'
+import { getEdgeTxSwapFromReportTxInfo } from '../../util/getEdgeTxSwapFromReportTxInfo'
 import { matchJson } from '../../util/matchJson'
 import { getMemoTitle } from '../../util/memoUtils'
+import { fetchGetTxInfo } from '../../util/reportsServer'
 import {
   convertCurrencyFromExchangeRates,
   convertNativeToExchange,
@@ -86,6 +90,37 @@ const TransactionDetailsComponent = (props: Props) => {
   const styles = getStyles(theme)
   const iconColor = useIconColor({ pluginId: currencyInfo.pluginId, tokenId })
 
+  const transactionSwapData =
+    convertActionToSwapData(account, transaction) ?? transaction.swapData
+
+  // Query for transaction info from reports server only if the transaction is
+  // a receive (we need to get potential swap data) or the transaction has
+  // swap data (we need to get the status)
+  const shouldShowTradeDetails =
+    !transaction.isSend || transactionSwapData != null
+  const { data: reportsTxInfo, isLoading: isReportsTxInfoLoading } = useQuery({
+    queryKey: ['txInfo', transaction.txid],
+    queryFn: async () => {
+      return await fetchGetTxInfo(transaction)
+    },
+    staleTime: query =>
+      // Only cache if the status has resolved, otherwise we'll always consider
+      // the data to be stale:
+      ['processing', 'pending', undefined].includes(query.state.data?.status)
+        ? 0 // No cache
+        : Infinity, // Cache forever
+    enabled: shouldShowTradeDetails,
+    retry: false
+  })
+
+  const swapDataFromReports = useMemo(
+    () =>
+      reportsTxInfo == null
+        ? null
+        : getEdgeTxSwapFromReportTxInfo(wallet, transaction, reportsTxInfo),
+    [reportsTxInfo, transaction, wallet]
+  )
+
   // Choose a default category based on metadata or the txAction
   const {
     action,
@@ -96,8 +131,7 @@ const TransactionDetailsComponent = (props: Props) => {
     savedData
   } = getTxActionDisplayInfo(transaction, account, wallet)
 
-  const swapData =
-    convertActionToSwapData(account, transaction) ?? transaction.swapData
+  const swapData = transactionSwapData ?? swapDataFromReports
 
   const thumbnailPath =
     useContactThumbnail(mergedData.name) ?? pluginIdIcons[iconPluginId ?? '']
@@ -546,7 +580,11 @@ const TransactionDetailsComponent = (props: Props) => {
           <SwapDetailsCard
             swapData={swapData}
             transaction={transaction}
-            sourceWallet={wallet}
+            sourceWallet={
+              swapData.payoutWalletId === wallet.id ? undefined : wallet
+            }
+            reportsTxInfo={reportsTxInfo}
+            isReportsTxInfoLoading={isReportsTxInfoLoading}
           />
         )}
       </EdgeAnim>
@@ -591,6 +629,7 @@ const TransactionDetailsComponent = (props: Props) => {
           )}
         />
       </EdgeAnim>
+
       <EdgeAnim enter={{ type: 'fadeInDown', distance: 140 }}>
         <ButtonsView
           layout="column"